requirements/base.txt
Django==1.9.2

# Basic Django extensions
django-extensions==1.6.1
django-braces==1.8.1
django-crispy-forms==1.6.0
django-model-utils==2.4
dogslow==0.9.7                       # For logging details about very slow requests
six==1.10.0                          # Dependency of django-extensions

# Basic Python libraries
Unipath==1.1          # Better path handling
wsgiref==0.1.2        # ???
pytz==2015.7          # Timezone data
ipython==4.1.1        # Prettier python prompt
gnureadline==6.3.3    # Dependency of ipython

# For generating docs - sphinx and dependencies
Sphinx==1.3.6
Jinja2==2.8
MarkupSafe==0.23
Pygments==2.1.2
docutils==0.12

# PEP8 validation with dependencies
flake8==2.5.4
mccabe==0.4.0
pyflakes==1.0.0
pep8==1.7.0

# Misc
python-dateutil==2.5.0
factory-boy==2.6.1
